Begin porting SelectionPane
